Dendrogram (#2)
* self-contained dendrogram generation and plot
* corrected walk_tree for multiple calls
