+++ examples/neurons/granule_cell.py
@@ -97,29 +97,9 @@ lb_axon = {
 
 ds.set_object_properties(n, axon_params=lb_axon)
 
-ds.simulate(5 * day)
+ds.simulate(7 * day)
+ds.plot.plot_dendrogram(n.axon, show=False)
 
 ds.io.save_to_swc("granule-cell.swc", gid=n)
 ds.plot.plot_neurons(show=True, subsample=50)
 
-import neurom as nm
-from neurom import viewer
-nrn = nm.load_neuron("granule-cell.swc")
-
-fig, _ = viewer.draw(nrn)
-
-for ax in fig.axes:
-    ax.set_title("")
-
-
-#~ tree = n[0].axon.get_tree()
-tree2 = n[0].axon.get_tree()
-print(tree2.neuron, tree2.neurite)
-
-import matplotlib.pyplot as plt
-plt.axis('off')
-fig.suptitle("")
-plt.tight_layout()
-plt.show()
-#~ tree.show_dendrogram()
-tree2.show_dendrogram()
