+++ examples/neurons/multipolar-cell.py
@@ -169,9 +169,8 @@ ds.set_object_properties(n, axon_params=end_branching, dendrites_params=end_bran
 ds.simulate(10*day)
 
 ds.plot.plot_recording(rec2, show=False)
+ds.plot.plot_dendrogram(n.axon, show=False)
 ds.plot.plot_neurons(show=True, scale_text=False)
 
 ds.io.save_to_swc("chandelier-cell.swc", gid=n, resolution=50)
 
-tree = n.axon.get_tree()
-tree.show_dendrogram()
