+++ examples/neurons/purkinje.py
@@ -48,7 +48,7 @@ neuron_params = {
     "axon_diameter": 3.*um,
     "position": (0., 0.)*um,
     "max_arbor_length": 20000.*um,
-    "diameter_eta_exp":1.,
+    "diameter_eta_exp": 20.,
     "diameter_ratio_avg": 1.,
 }
 
@@ -66,8 +66,8 @@ dendrite_params = {
     "use_van_pelt": True,
 
     "persistence_length": 100.0 * um,
-    "taper_rate": 1./80.,
-    "diameter_fraction_lb": 0.5,
+    "taper_rate": 1./100.,
+    "diameter_fraction_lb": 0.7,
 
     # SFR parameters
     "somatropic_scale": 100.*um,
@@ -101,7 +101,8 @@ if __name__ == '__main__':
     rec = ds.create_recorders(neuron, "num_growth_cones", levels="neuron")
 
     ds.simulate(2*hour)
-    ds.plot.plot_neurons()
+    # ds.plot.plot_dendrogram(neuron.dendrites["dendrite_1"], show=False)
+    # ds.plot.plot_neurons()
 
     neuron.dendrites["dendrite_1"].set_properties({
         "B": 6.*cpm, "T": 5.*hour,
@@ -109,7 +110,8 @@ if __name__ == '__main__':
     })
 
     ds.simulate(15*hour)
-    ds.plot.plot_neurons()
+    # ds.plot.plot_dendrogram(neuron.dendrites["dendrite_1"], show=False)
+    # ds.plot.plot_neurons()
 
     neuron.set_properties(dendrites_params={
         "use_van_pelt": False, "use_uniform_branching": True,
@@ -119,6 +121,8 @@ if __name__ == '__main__':
     })
 
     ds.simulate(6*day)
+    ds.plot.plot_dendrogram(neuron.dendrites["dendrite_1"],
+                            ignore_diameter=True, aspect_ratio=0.2, show=False)
     ds.plot.plot_neurons()
 
     neuron.set_properties(dendrites_params={
@@ -129,5 +133,6 @@ if __name__ == '__main__':
     })
 
     ds.simulate(20.*day)
+    ds.plot.plot_dendrogram(neuron.dendrites["dendrite_1"], show=False)
     ds.plot.plot_neurons()
     ds.plot.plot_recording(rec)
