+++ examples/neurons/pyramidal_cell.py
@@ -59,7 +59,7 @@ axon_params = {
     "growth_cone_model": gc_model,
 
     # Steering parameters
-    "sensing_angle": 90.*deg,
+    "sensing_angle": 80.*deg,
     # "self_avoidance_factor": 0.,
     # "self_avoidance_scale": 20.*um,
     "somatropic_scale": 70.*um,
@@ -67,7 +67,7 @@ axon_params = {
 
     #"filopodia_wall_affinity": 0.05,
     "filopodia_finger_length": 20.*um,
-    "filopodia_min_number": 30,    
+    "filopodia_min_number": 30,
 
     # extension parameters
     "persistence_length": 500.*um,
@@ -85,7 +85,7 @@ axon_params = {
 dend_params = {
     "growth_cone_model": gc_model,
     # Steering parameters
-    "sensing_angle": 90.*deg,
+    "sensing_angle": 80.*deg,
 
     "somatropic_mode": "window",
     # "somatropic_factor": 100.,
@@ -136,12 +136,8 @@ rec = ds.create_recorders(n, "num_growth_cones")
 
 ds.simulate(10*day)
 
-print(ds.get_kernel_status('time'))
 
-recording = ds.get_recording(rec, record_format="compact")
-print(recording)
-
-ds.plot.plot_neurons(mode="mixed", show=True)
+# ~ ds.plot.plot_neurons(mode="mixed", show=True)
 
 # second development phase : with lateral branching
 
@@ -175,6 +171,7 @@ ds.set_object_properties(n, dendrites_params=dend_params, axon_params=lb_axon)
 
 ds.simulate(7*day)
 
+ds.plot.plot_dendrogram(n.axon, show=False)
 ds.plot.plot_neurons(mode="mixed", show=True)
 
 # Now a third step in development
@@ -195,4 +192,5 @@ dend_params = {
 ds.set_object_properties(n, dendrites_params=dend_params, axon_params=vp_axon)
 ds.simulate(20*day)
 
+ds.plot.plot_dendrogram(n.axon, show_node_id=True, show=False)
 ds.plot.plot_neurons(scale_text=False)
