+++ src/pymodule/dense/elements.py
@@ -58,13 +58,13 @@ class Neuron(object):
 
     def __ge__(self, other):
         return int(self) >= int(other)
-    
+
     def __str__(self):
         return str(self.__gid)
-    
+
     def __repr__(self):
         return "Neuron<{}>".format(self.__gid)
-    
+
     def _repr_pretty_(self, p, cycle):
         p.text("Neuron({})".format(self.__gid))
 
@@ -129,7 +129,7 @@ class Neuron(object):
     def total_length(self):
         ''' Total arbor length of the neuron '''
         return _pg.get_object_state(self, variable="length")
-    
+
     def create_neurites(self, num_neurites=1, params=None, angles=None,
                         neurite_types=None, names=None):
         '''
@@ -150,7 +150,7 @@ class Neuron(object):
             will be dendrites.
         names : str or list, optional (default: "axon" and "dendrite_X")
             Names of the created neurites.
-        
+
         See also
         --------
         :func:`~dense.create_neurites`.
@@ -205,7 +205,7 @@ class Neuron(object):
         status : variable
             Properties of the objects' status: a single value if
             `property_name` was specified, the full status ``dict`` otherwise.
-        
+
         See also
         --------
         :func:`~dense.get_object_properties`,
@@ -348,7 +348,7 @@ class Neuron(object):
                         # set as next parent
                         p_segment = n_segment
                         parent    = neuroml.SegmentParent(segments=p_segment.id)
-                        seg_id += 1 
+                        seg_id += 1
 
                         neurites_segments.append(n_segment)
 
@@ -369,7 +369,7 @@ class Neuron(object):
         # write
         if write:
             doc = neuroml.NeuroMLDocument(id=filename)
-            doc.cells.append(cell)        
+            doc.cells.append(cell)
             writers.NeuroMLWriter.write(doc, filename)
 
         return cell
@@ -402,7 +402,7 @@ class Neurite(object):
 
     def __str__(self):
         return self.__name
-    
+
     def __repr__(self):
         if self._parent is None:
             return "Neurite<{} at {}>".format(str(self), id(self))
@@ -433,49 +433,6 @@ class Neurite(object):
 
     def get_tree(self):
         return _pg._get_tree(self._parent, str(self))
-
-    def get_properties(self, property_name=None):
-        '''
-        Get the neurite's properties.
-
-        Parameters
-        ----------
-        property_name : str, optional (default: None)
-            Name of the property that should be queried. By default, the full
-            dictionary is returned.
-
-        Returns
-        -------
-        status : variable
-            Properties of the neurite: a single value if
-            `property_name` was specified, the full status ``dict`` otherwise.
-        
-        See also
-        --------
-        :func:`~dense.get_object_properties`,
-        :func:`~dense.elements.Neurite.set_properties`,
-        :func:`~dense.elements.Neuron.get_properties`.
-        '''
-        return _pg.get_object_properties(self._parent, property_name=property_name,
-                                         neurite=self)
-
-    def set_properties(self, params):
-        '''
-        Update the neurite parameters using the entries contained in `params`.
-
-        Parameters
-        ----------
-        params : dict
-            New neurite parameters.
-
-        See also
-        --------
-        :func:`~dense.set_object_properties`,
-        :func:`~dense.elements.Neurite.get_properties`,
-        :func:`~dense.elements.Neuron.set_properties`.
-        '''
-        return _pg.set_neurite_properties(self._parent, self, params=params)
-
 
     @property
     def name(self):
@@ -578,7 +535,7 @@ class Neurite(object):
         status : variable
             Properties of the objects' status: a single value if
             `property_name` was specified, the full status ``dict`` otherwise.
-        
+
         See also
         --------
         :func:`~dense.get_object_properties`,
@@ -682,13 +639,24 @@ class Node(int):
         self.position       = pos
         self.dist_to_parent = dist_to_parent
         self.children       = []
-        self.parent         = (tree.get(int(parent), parent)
+        self.parent         = (tree.get(parent, parent)
                                if parent is not None else None)
 
     def add_child(self, child):
         self.children.append(child)
         self._tree[int(child)] = child
         child.parent           = self
+
+    def distance_to_soma(self):
+        dts = self.dist_to_parent
+
+        node = self._tree.get(self.parent, None)
+
+        while node is not None:
+            dts += node.dist_to_parent
+            node = self._tree.get(node.parent, None)
+
+        return dts
 
 
 class Tree(dict):
@@ -698,6 +666,7 @@ class Tree(dict):
         self._neuron    = neuron
         self._neurite   = neurite
         self._root      = None
+        self._tips      = set()
         self._tips_set  = False
 
     @property
@@ -718,17 +687,17 @@ class Tree(dict):
         return self._tips
 
     def __setitem__(self, key, value):
-        super(Tree, self).__setitem__(key, value)
-        if value.parent is None or value.parent:
+        super(Tree, self).__setitem__(int(key), value)
+        if value.parent is None or value.parent == value:
             self._root = value
         value._tree    = self
         self._tips_set = False
 
     def update_tips(self):
-        self._tips = []
-        for key, val in self.items():
+        self._tips = set()
+        for val in self.values():
             if not val.children:
-                self._tips.append(val)
+                self._tips.add(val)
         self._tips_set = True
 
     def neurom_tree(self):
@@ -757,60 +726,34 @@ class Tree(dict):
 
         return neuron
 
-    def show_dendrogram(self):
+    def show_dendrogram(self, **kwargs):
         '''
-        Make and display the dendrogram using ETE3
+        Make and display the dendrogram.
 
-        Returns
-        -------
-        t, ts : ete3.Tree, ete3.TreeStyle
+        See also
+        --------
+        Plotting arguments are the same as :func:`dense.plot.plot_dendrogram`.
         '''
-        from ete3 import Tree as Ete3Tree
-        from ete3 import TreeStyle, NodeStyle
-
-        # make the tree from the root
-        t      = Ete3Tree(dist=0, name=int(self._root))
-        ns = NodeStyle()
-        ns["hz_line_width"] = self._root.diameter
-        ns["size"]          = 0
-        t.set_style(ns)
-
-        queue  = _deque(self._root.children)
-        edict  = {int(self._root): t}
-
-        while queue:
-            node   = queue.popleft()
-            parent = edict[node.parent]
-            enode  = parent.add_child(name=int(node), dist=node.dist_to_parent)
-
-            ns = NodeStyle()
-            # ~ ns["vt_line_width"] = node.diameter
-            ns["hz_line_width"] = node.diameter
-            ns["size"]          = 0
-            enode.set_style(ns)
-
-            edict[node] = enode
-            queue.extend(node.children)
-
-        # set style
-        ts = TreeStyle()
-        ts.show_leaf_name = False
-        ts.branch_vertical_margin = self._root.diameter
-
-        # show
-        t.show(tree_style=ts)
-
-        return t, ts
+        from .plot import plot_dendrogram
+        plot_dendrogram(self, **kwargs)
 
     def _cleanup(self):
+        '''
+        This step is necessary to use the Tree properly since it converts
+        parents from int to Node.
+        '''
         for val in self.values():
             val.children = []
+
         for key, val in self.items():
             if val.parent is None or val.parent == val:
                 self._root = val
                 val.parent = None
+                self[val] = val
             else:
-                self[int(val.parent)].children.append(val)
+                self[val.parent].children.append(val)
+                val.parent = self[val.parent]
+
         self.update_tips()
 
 
@@ -975,7 +918,7 @@ class Population(list):
         '''
         add population
         '''
-        from .io import GetSWCStructure as _get_swc_struct
+        from .io.data_swc import GetSWCStructure as _get_swc_struct
 
         for neuron in neurons:
             gid = neurons[neuron]['gid']
