+++ src/pymodule/dense/plot/__init__.py
@@ -19,40 +19,18 @@
 # You should have received a copy of the GNU General Public License
 # along with DeNSE. If not, see <http://www.gnu.org/licenses/>.
 
+"""
+Plotting module
+===============
+"""
 
-""" IO functions """
+from .plot_recording import plot_recording
+from .plot_structures import plot_neurons, plot_environment, plot_dendrogram
 
-import hashlib
-import json
-
-from .data_swc import load_swc, save_to_swc
-from .dataIO import (save_json_info, save_to_neuroml)
 
 __all__ = [
-    "generate_hash_id",
-    "load_swc",
-    "save_json_info",
-    "save_to_neuroml",
-    "save_to_swc",
+    "plot_dendrogram",
+    "plot_environment",
+    "plot_neurons",
+    "plot_recording",
 ]
-
-
-# ---------- #
-# Hash tools #
-# ---------- #
-
-def generate_hash_id(*args):
-    '''
-    Return the hash ID of an experiment.
-    '''
-    experiment_dict = {}
-    for num, dict_ in enumerate(args):
-        experiment_dict[num] = dict_
-    return _hash_dict(experiment_dict)
-
-
-def _hash_dict(_dict):
-    sha = hashlib.sha1()
-    sha.update(str(json.dumps(_dict, sort_keys =True)).encode('utf-8'))
-    hash_name = sha.hexdigest()
-    return hash_name[:16]
